+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+-   Word Count
+-   Full formatted paste features, `ctrl`+`shift`+`v`
+    -   Reformats some Unicode text for LaTeX
+    -   Table cells are turned into a `tabular`
+    -   Pasting the location of a `.csv` file pastes a table with the contents
+    -   Paste an image from the clipboard (as in 0.1.0)
+    -   Formatted paste the path to a csv (adds tabular) or image file (links to file)
+
 ## [0.1.0]
 
 ### Added